+++ anchor-contributor/Anchor.toml
@@ -1,7 +1,7 @@
 [features]
 seeds = false
 [programs.localnet]
-anchor_contributor = "Efzc4SLs1ZdTPRq95oWxdMUr9XiX5M14HABwHpvrc9Fm"
+anchor_contributor = "iccoWYbaJAr8dC1MHGLe6G8Vtsu1dEoWS3yZpj9zKyC"
 
 [programs.devnet]
 anchor_contributor = "iccoWYbaJAr8dC1MHGLe6G8Vtsu1dEoWS3yZpj9zKyC"
