+++ sdk/package.json
@@ -10,11 +10,10 @@
   ],
   "repository": "https://github.com/certusone/wormhole-icco/tree/main/sdk",
   "scripts": {
-    "clean": "rm -rf node_modules evm-contracts src/evm/ethers-contracts src/solana/types",
-    "build:evm": "npm run build --prefix ../ethereum && cp -r ../ethereum/build/contracts ./evm-contracts && typechain --target=ethers-v5 --out-dir=src/ethers-contracts evm-contracts/*.json",
-    "cp:solana": "cp -r ../anchor-contributor/target/types ./src/solana/ && cp -r ../anchor-contributor/target/idl ./solana-idl",
-    "build": "npm run build:evm && npm run cp:solana",
-    "testnet-integration-test": "ts-mocha src/testnet/run_*.ts -t 1000000",
+    "clean": "rm -rf node_modules evm-contracts target src/ethers-contracts src/anchor",
+    "build": "bash scripts/copy_ethereum.sh && bash scripts/copy_anchor.sh",
+    "testnet-integration-test": "ts-mocha src/testnet/run_testnet_sale.ts -t 1000000",
+    "testnet-solana-test": "ts-mocha src/testnet/run_solana_sales.ts -t 1000000",
     "format": "echo \"disabled: prettier --write \"src/**/*.ts\"\"",
     "lint": "tslint -p tsconfig.json",
     "prepublishOnly": "echo \"disabled: npm test && npm run lint\"",
